fix: resolve type mismatches between Float and Double in rating components
- Change InteractiveRatingDisplay to use Float? instead of Double?
- Update ReviewDetailsSheet to use Float for rating parameter
- Convert Float to Double when passing to CompactStarRating
- Fix literal values (0.0 -> 0f) to match Float types
- Addresses compilation errors from install-output.log

+++ feature/playlist/src/main/java/com/deadarchive/feature/playlist/components/ReviewDetailsSheet.kt
@@ -22,7 +22,7 @@ import com.deadarchive.feature.playlist.data.Review
 @Composable
 fun ReviewDetailsSheet(
     recordingTitle: String,
-    rating: Double,
+    rating: Float,
     reviewCount: Int,
     ratingDistribution: Map<Int, Int>? = null,
     reviews: List<Review> = emptyList(),
@@ -173,7 +173,7 @@ fun ReviewDetailsSheet(
 
 @Composable
 private fun RatingSummaryCard(
-    rating: Double,
+    rating: Float,
     reviewCount: Int,
     ratingDistribution: Map<Int, Int>?
 ) {
@@ -196,7 +196,7 @@ private fun RatingSummaryCard(
             )
             
             CompactStarRating(
-                rating = rating,
+                rating = rating.toDouble(),
                 confidence = null,
                 starSize = IconResources.Size.MEDIUM
             )
